refactor, a lot
