+++ src/app.ts
@@ -1,20 +1,20 @@
-import { ConfigManager } from "./configManager";
 import * as net from "net";
 import * as repl from "repl";
+import { PatchServer } from "./services/PatchAndShard/patchServer";
+import { WebServer } from "./services/AuthLogin/WebServer";
+import { Server } from "./server";
+import { ConfigManager } from "./services/shared/configManager";
 
 // get instance of config
 const config = new ConfigManager().getConfig();
 
 // MCOS Monolith
-import { Server } from "./server";
 const server = new Server();
 server.start();
 // MCOS AuthLogin and Shard
-import { WebServer } from "./services/AuthLogin/WebServer";
 const AuthLogin = new WebServer();
 AuthLogin.start();
 // MCOS PatchAndShard
-import { PatchServer } from "./services/PatchAndShard/patchServer";
 const patchAndShardServer = new PatchServer();
 patchAndShardServer.start();
