+++ src/services/AuthLogin/WebServer.test.ts
@@ -1,15 +1,11 @@
 import { WebServer } from "./WebServer";
 import * as fs from "fs";
-import { ConfigManager } from "../../configManager";
-
-// get instance of config
-const config = new ConfigManager().getConfig();
 
 describe("WebServer", () => {
   test("can generate registry file", () => {
     const webServer = new WebServer();
     const staticRegistry = fs.readFileSync(
-      config.serverConfig.registryFilename
+      webServer.config.serverConfig.registryFilename
     );
     const dynamicRegistry = webServer._handleGetRegistry();
     expect(dynamicRegistry).toEqual(staticRegistry.toString("UTF-8"));
