+++ src/services/MCServer/TCPManager.ts
@@ -9,16 +9,16 @@ import * as assert from "assert";
 import { Socket } from "net";
 import { ConnectionObj } from "../../ConnectionObj";
 import { IRawPacket } from "../shared/interfaces/IRawPacket";
-import { LobbyServer } from "../../LobbyServer/LobbyServer";
-import { MCOTServer } from "../../MCOTS/MCOTServer";
-import { ClientConnectMsg } from "../shared/messageTypes/ClientConnectMsg";
+import { LobbyServer } from "./LobbyServer/LobbyServer";
+import { MCOTServer } from "./MCOTS/MCOTServer";
+import { ClientConnectMsg } from "./ClientConnectMsg";
 import { GenericReplyMsg } from "../shared/messageTypes/GenericReplyMsg";
 import { GenericRequestMsg } from "../shared/messageTypes/GenericRequestMsg";
 import { MessageNode } from "../shared/messageTypes/MessageNode";
-import { StockCar } from "../shared/messageTypes/StockCar";
-import { StockCarInfoMsg } from "../shared/messageTypes/StockCarInfoMsg";
-import { DatabaseManager } from "../../databaseManager";
-import { Logger } from "../../loggerManager";
+import { StockCar } from "./messageTypes/StockCar";
+import { StockCarInfoMsg } from "./messageTypes/StockCarInfoMsg";
+import { DatabaseManager } from "../shared/databaseManager";
+import { Logger } from "../shared/loggerManager";
 
 const logger = new Logger().getLogger("TCPManager");
 const lobbyServer = new LobbyServer();
