--- src/services/MCServer/loggerManager.ts
@@ -1,20 +1,0 @@
-import * as bunyan from "bunyan";
-const BunyanToGelfStream = require("bunyan-gelf");
-
-export class Logger {
-  public getLogger(service: string) {
-    const logger = bunyan
-      .createLogger({ name: "mcoServer" })
-      .child({ module: service });
-    logger.addStream({
-      type: "raw",
-      stream: new BunyanToGelfStream({
-        host: "graylog",
-        port: 12201,
-      }),
-    });
-    logger.addSerializers(bunyan.stdSerializers);
-
-    return logger;
-  }
-}
