+++ src/services/MCServer/messageTypes/StockCarInfoMsg.test.ts
