+++ src/services/PatchAndShard/patchServer.test.ts
@@ -1,6 +1,6 @@
 import { PatchServer } from "./patchServer";
 import * as fs from "fs";
-import { ConfigManager } from "../../configManager";
+import { ConfigManager } from "../shared/configManager";
 
 // get instance of config
 const config = new ConfigManager().getConfig();
