+++ src/services/shared/databaseManager.ts
